requirements.txt
Django
celery
redis
requests
pillow
django-celery-results
python-dotenv
gevent
